+++ ibc-clients/ics07-tendermint/types/src/consensus_state.rs
@@ -147,28 +147,3 @@ impl From<Header> for ConsensusState {
         Self::from(header.signed_header.header)
     }
 }
-
-#[cfg(all(test, feature = "serde"))]
-mod tests {
-    use tendermint_rpc::endpoint::abci_query::AbciQuery;
-
-    use crate::serde_tests::test_serialization_roundtrip;
-
-    #[test]
-    fn serialization_roundtrip_no_proof() {
-        let json_data = include_str!(concat!(
-            env!("CARGO_MANIFEST_DIR"),
-            "/../../../tests-integration/tests/data/json/consensus_state.json"
-        ));
-        test_serialization_roundtrip::<AbciQuery>(json_data);
-    }
-
-    #[test]
-    fn serialization_roundtrip_with_proof() {
-        let json_data = include_str!(concat!(
-            env!("CARGO_MANIFEST_DIR"),
-            "/../../../tests-integration/tests/data/json/consensus_state_proof.json"
-        ));
-        test_serialization_roundtrip::<AbciQuery>(json_data);
-    }
-}
