+++ ibc-testkit/src/fixtures/clients/tendermint.rs
@@ -15,6 +15,8 @@ use ibc::core::client::types::Height;
 use ibc::core::commitment_types::specs::ProofSpecs;
 use ibc::core::host::types::identifiers::ChainId;
 use ibc::core::primitives::prelude::*;
+#[cfg(feature = "serde")]
+use serde::{de::DeserializeOwned, Serialize};
 use tendermint::block::Header as TmHeader;
 use typed_builder::TypedBuilder;
 
@@ -147,7 +149,7 @@ pub fn dummy_ics07_header() -> Header {
     // Build a SignedHeader from a JSON file.
     let shdr = serde_json::from_str::<SignedHeader>(include_str!(concat!(
         env!("CARGO_MANIFEST_DIR"),
-        "/../tests-integration/tests/data/json/signed_header.json"
+        "/src/data/json/signed_header.json"
     )))
     .expect("Never fails");
 
@@ -172,9 +174,28 @@ pub fn dummy_ics07_header() -> Header {
     }
 }
 
+#[cfg(feature = "serde")]
+pub fn test_serialization_roundtrip<T>(json_data: &str)
+where
+    T: core::fmt::Debug + PartialEq + Serialize + DeserializeOwned,
+{
+    let parsed0 = serde_json::from_str::<T>(json_data);
+    assert!(parsed0.is_ok());
+    let parsed0 = parsed0.expect("should not fail");
+
+    let serialized = serde_json::to_string(&parsed0);
+    assert!(serialized.is_ok());
+    let serialized = serialized.expect("should not fail");
+
+    let parsed1 = serde_json::from_str::<T>(&serialized);
+    assert!(parsed1.is_ok());
+    let parsed1 = parsed1.expect("should not fail");
+
+    assert_eq!(parsed0, parsed1);
+}
+
 #[cfg(all(test, feature = "serde"))]
 mod tests {
-
     use ibc::primitives::proto::Any;
     use rstest::rstest;
