+++ tests-integration/tests/clients/ics07_tendermint/mod.rs
@@ -1,0 +1,2 @@
+pub mod client_state;
+pub mod consensus_state;
