+++ tests-integration/tests/clients/mod.rs
@@ -1,0 +1,1 @@
+pub mod ics07_tendermint;
