+++ tests-integration/tests/mod.rs
@@ -7,5 +7,6 @@
     rust_2018_idioms
 )]
 pub mod applications;
+pub mod clients;
 pub mod core;
 pub mod cosmwasm;
